+++ test/unit/layeredmaterial.js
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import LayerUpdateState from '../../src/Core/Layer/LayerUpdateState';
-import ColorTextureProcessing from '../../src/Process/ColorTextureProcessing';
+import TextureProcessing from '../../src/Process/TextureProcessing';
 
 describe('material state vs layer state', function () {
     let opacity;
@@ -26,21 +26,21 @@ describe('material state vs layer state', function () {
 
     it('should correctly initialize opacity & visibility', () => {
         node.layerUpdateState.test.failure(new Date());
-        ColorTextureProcessing.updateLayerElement(null, layer, node);
+        TextureProcessing.Color.updateLayerElement(null, layer, node);
         assert.equal(opacity, layer.opacity);
         assert.equal(visible, layer.visible);
     });
     it('should update material opacity & visibility', () => {
         layer.opacity = 0.5;
         layer.visible = false;
-        ColorTextureProcessing.updateLayerElement(null, layer, node);
+        TextureProcessing.Color.updateLayerElement(null, layer, node);
         assert.equal(opacity, layer.opacity);
         assert.equal(visible, layer.visible);
     });
     it('should update material opacity & visibility even if layer is cannot be updated', () => {
         node.layerUpdateState.test.noMoreUpdatePossible();
         layer.opacity = 0.75;
-        ColorTextureProcessing.updateLayerElement(null, layer, node);
+        TextureProcessing.Color.updateLayerElement(null, layer, node);
         assert.equal(opacity, layer.opacity);
         assert.equal(visible, layer.visible);
     });
